requirements.txt
streamlit==1.31.0
openai==1.10.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0
python-dotenv==1.0.0
numpy==1.26.0
requests==2.31.0
